data streamer

+++ src/task/task.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma.service';
 import { Task } from '@prisma/client';
 import { CreateTaskDto } from './taskDto';
@@ -7,12 +7,16 @@ import { TaskStatus } from './TaskStatus';
 
 @Injectable()
 export class TaskService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(@Inject(PrismaService) private readonly prisma: PrismaService) {}
 
   async getTaskById(id: number): Promise<Task> {
-    return this.prisma.task.findFirstOrThrow({
+    const task = await this.prisma.task.findFirst({
       where: { id },
     });
+    if (!task) {
+      throw new NotFoundException('Task not found');
+    }
+    return task;
   }
 
   async createTask(body: CreateTaskDto, userId: number) {
@@ -22,19 +26,32 @@ export class TaskService {
         userId,
         status: TaskStatus.TODO,
         dueDateSet: body.dueDate && true,
+        dueDate: body.dueDate && new Date(body.dueDate),
       },
     });
   }
   async updateTask(id: number, body: Partial<Task>) {
+    const task = await this.getTaskById(id);
+    if (!task) {
+      throw new NotFoundException('Task not found');
+    }
     return this.prisma.task.update({
       where: {
         id,
       },
-      data: { ...body },
+      data: {
+        ...body,
+        dueDate: body.dueDate ? new Date(body.dueDate) : null,
+        dueDateSet: !!body.dueDate,
+      },
     });
   }
 
   async deleteTask(id: number) {
+    const task = await this.getTaskById(id);
+    if (!task) {
+      throw new NotFoundException('Task not found');
+    }
     return this.prisma.task.delete({
       where: {
         id,
@@ -65,7 +82,7 @@ export class TaskService {
           }),
         ...(query.status && { status: query.status }),
       },
-      skip: query.skip(),
+      skip: query.skip,
       take: query.take,
       orderBy: {
         id: query.order,
